Track RB velocity in case collision sets it to zero
Should fix issue when player is accelerating into wall and then switches directions. Note that current method exhibits the behaviour that when the player is at a wall and directional input is poining at it, curentHorizontalSpeed is non-zero. I presume this is non-zero during computeHorizontalSpeed() only for a sigle update step before the speed is set to zero once more at the start of next computeHorizontalSpeed().