WIP: Add vertical movement
Very broken. Vertical movement currenly clips through everything. I suspect that this is because of applying non-zero vertical speed to a colliding RB? Not sure. There is also likely some shenaniganz with the sign of the vertical speed. The implemented method tries to emulate the one seen in the resouces sent to the Resources slack channel.